% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{file_bom_type}
\alias{file_bom_type}
\title{Get BOM type (file)}
\usage{
file_bom_type(path)
}
\arguments{
\item{path}{path to a file}
}
\description{
Get BOM type (file)
}
\examples{
file_bom_type(system.file("examples", "stop_times.txt", package="bom"))
}

