% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{file_has_bom}
\alias{file_has_bom}
\title{Test if a file has a BOM}
\usage{
file_has_bom(path)
}
\arguments{
\item{path}{path to a file}
}
\description{
Test if a file has a BOM
}
\examples{
file_has_bom(system.file("examples", "stops.txt", package="bom"))
file_has_bom(system.file("examples", "stop_times.txt", package="bom"))
}

