% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raw_bom_type}
\alias{raw_bom_type}
\title{Get BOM type (raw vector)}
\usage{
raw_bom_type(x)
}
\arguments{
\item{x}{raw vector}
}
\description{
Get BOM type (raw vector)
}
\examples{
raw_bom_type(readBin(system.file("examples", "stop_times.txt", package="bom"), "raw", 4))
}

